requirements.txt
Pillow==10.2.0
numpy==1.26.4
pytest==8.1.1 